pyproject.toml
[tool.poetry]
name = "crypto-notif"
version = "0.1.0"
description = ""
authors = ["vanilara <vanilara@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
sqlalchemy = "^2.0.36"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
aiohttp = "3.10"
pytest-asyncio = "^0.24.0"
pycryptodome = "^3.21.0"
pytest-dotenv = "^0.5.2"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
greenlet = "^3.1.1"
web3 = "^7.6.0"
eth-utils = "^5.1.0"
pydantic-settings = "^2.6.1"
pydantic = "2.9"
aiogram = "^3.15.0"
base58 = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
